added edit functionality

+++ controllers/order.ts
@@ -78,9 +78,20 @@ export const updateOrder = async (req: Request, res: Response) => {
 export const getOrders = async (req: Request, res: Response) => {
   try {
     const { userId } = req.query;
+    const user = await prisma.user.findUnique({
+      where: {
+        id: userId as string,
+      },
+    });
+
     const where: any = {};
-    if (userId) {
+    if (user?.role === "SALESMAN") {
       where.userId = userId as string;
+    }
+    if (user?.role === "DISPATCHER") {
+      where.status = {
+        in: ["completed", "approved"],
+      };
     }
     const orders = await prisma.order.findMany({
       where,
@@ -124,13 +135,27 @@ export const getOrder = async (req: Request, res: Response) => {
 
 export const deleteOrder = async (req: Request, res: Response) => {
   try {
-    const { id } = req.params;
-    const order = await prisma.order.delete({
-      where: {
-        id: id,
-      },
-    });
-    res.status(200).json({ success: true, order });
+    const { id, from, to } = req.query;
+    if (id) {
+      const order = await prisma.order.delete({
+        where: {
+          id: id as string,
+        },
+      });
+    }
+    if (from && to) {
+      const orders = await prisma.order.deleteMany({
+        where: {
+          createdAt: {
+            gte: new Date(from as string),
+            lte: new Date(to as string),
+          },
+        },
+      });
+    }
+    console.log(req.query, new Date(from as string), new Date(to as string));
+
+    res.status(200).json({ success: true });
   } catch (error) {
     res.status(400).json({ success: false, error });
   }
